NGSTACK-842 restructure bundle
