+++ tests/bundle/Integration/ContentTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Netgen\IbexaScheduledVisibility\Tests\Integration;
 
-use Netgen\Bundle\IbexaScheduledVisibilityBundle\ScheduledVisibility\Handler\Content;
+use Netgen\Bundle\IbexaScheduledVisibilityBundle\Core\VisibilityHandler\Content;
 
 final class ContentTest extends BaseTest
 {
